minor refactor and fixed unit tests
