+++ polaris-ui/src/app/features/cases/hooks/utils/redactionUtils.test.ts
@@ -4,8 +4,8 @@ import {
 } from "./redactionUtils";
 
 describe("redactionUtils", () => {
-  xdescribe("getNormalizedRedactionHighlights", () => {
-    it("should return correct normalized highlight data for a smaller canvas height", () => {
+  describe("getNormalizedRedactionHighlights", () => {
+    it("Should not normalize if there is no resize (height remains same)", () => {
       const redactionHighlights = [
         {
           highlightType: "area" as "area" | "linear",
@@ -28,20 +28,17 @@ describe("redactionUtils", () => {
             "This is an area redaction and redacted content is unavailable",
           type: "redaction" as "redaction" | "search",
         },
-      ];
-
-      const expectedResult = [
         {
           highlightType: "area" as "area" | "linear",
-          id: "1707153788147",
+          id: "1707153788148",
           position: {
             boundingRect: {
-              width: 1142.86,
-              height: 800,
-              x1: 114.29,
-              y1: 114.29,
-              x2: 171.43,
-              y2: 171.43,
+              height: 700,
+              width: 1000,
+              x1: 200,
+              y1: 200,
+              x2: 250,
+              y2: 250,
             },
             rects: [],
             pageNumber: 1,
@@ -52,141 +49,114 @@ describe("redactionUtils", () => {
             "This is an area redaction and redacted content is unavailable",
           type: "redaction" as "redaction" | "search",
         },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 800,
+              width: 1100,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
       ];
+
       expect(getNormalizedRedactionHighlights(redactionHighlights)).toEqual(
-        expectedResult
+        redactionHighlights
       );
     });
-    it("should return correct normalized highlight data for a bigger canvas height", () => {
+    it("should return correct normalized highlight data for resized page redactions", () => {
       const redactionHighlights = [
         {
           highlightType: "area" as "area" | "linear",
           id: "1707153788147",
           position: {
             boundingRect: {
+              height: 700,
               width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [],
+            pageNumber: 1,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788148",
+          position: {
+            boundingRect: {
               height: 900,
-              x1: 100,
-              y1: 100,
-              x2: 150,
-              y2: 150,
-            },
-            rects: [],
-            pageNumber: 1,
-          },
-          redactionAddedOrder: 0,
-          redactionType: { id: "1", name: "Named individual" },
-          textContent:
-            "This is an area redaction and redacted content is unavailable",
-          type: "redaction" as "redaction" | "search",
-        },
-      ];
-
-      const expectedResult = [
-        {
-          highlightType: "area" as "area" | "linear",
-          id: "1707153788147",
-          position: {
-            boundingRect: {
-              width: 888.89,
-              height: 800,
-              x1: 88.89,
-              y1: 88.89,
-              x2: 133.33,
-              y2: 133.33,
-            },
-            rects: [],
-            pageNumber: 1,
-          },
-          redactionAddedOrder: 0,
-          redactionType: { id: "1", name: "Named individual" },
-          textContent:
-            "This is an area redaction and redacted content is unavailable",
-          type: "redaction" as "redaction" | "search",
-        },
-      ];
-      expect(getNormalizedRedactionHighlights(redactionHighlights)).toEqual(
-        expectedResult
-      );
-    });
-
-    it("should return correct normalized highlight data for multiple redactionHighlights ", () => {
-      const redactionHighlights = [
-        {
-          highlightType: "area" as "area" | "linear",
-          id: "1707153788147",
-          position: {
-            boundingRect: {
-              width: 700,
-              height: 600,
-              x1: 10,
-              y1: 10,
-              x2: 20,
-              y2: 30,
-            },
-            rects: [],
-            pageNumber: 1,
-          },
-          redactionAddedOrder: 0,
-          redactionType: { id: "1", name: "Named individual" },
-          textContent:
-            "This is an area redaction and redacted content is unavailable",
-          type: "redaction" as "redaction" | "search",
-        },
-        {
-          highlightType: "area" as "area" | "linear",
-          id: "1707153788147",
-          position: {
-            boundingRect: {
-              width: 900,
-              height: 1000,
-              x1: 10,
-              y1: 10,
-              x2: 20,
-              y2: 30,
-            },
-            rects: [],
-            pageNumber: 1,
-          },
-          redactionAddedOrder: 1,
-          redactionType: { id: "1", name: "Named individual" },
-          textContent:
-            "This is an area redaction and redacted content is unavailable",
-          type: "redaction" as "redaction" | "search",
-        },
-        {
-          highlightType: "linear" as "area" | "linear",
-          id: "1707153788147",
-          position: {
-            boundingRect: {
               width: 1000,
-              height: 700,
+              x1: 200,
+              y1: 200,
+              x2: 250,
+              y2: 250,
+            },
+            rects: [],
+            pageNumber: 1,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 800,
+              width: 1100,
               x1: 100,
               y1: 100,
               x2: 150,
               y2: 150,
             },
-            rects: [
-              {
-                width: 1000,
-                height: 700,
-                x1: 100,
-                y1: 100,
-                x2: 150,
-                y2: 150,
-              },
-              {
-                width: 1000,
-                height: 700,
-                x1: 110,
-                y1: 110,
-                x2: 160,
-                y2: 160,
-              },
-            ],
-            pageNumber: 1,
+            rects: [],
+            pageNumber: 2,
           },
-          redactionAddedOrder: 2,
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 1000,
+              width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
           redactionType: { id: "1", name: "Named individual" },
           textContent:
             "This is an area redaction and redacted content is unavailable",
@@ -200,12 +170,12 @@ describe("redactionUtils", () => {
           id: "1707153788147",
           position: {
             boundingRect: {
-              width: 1073.33,
-              height: 920,
-              x1: 15.33,
-              y1: 15.33,
-              x2: 30.67,
-              y2: 46,
+              height: 700,
+              width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
             },
             rects: [],
             pageNumber: 1,
@@ -218,58 +188,62 @@ describe("redactionUtils", () => {
         },
         {
           highlightType: "area" as "area" | "linear",
-          id: "1707153788147",
+          id: "1707153788148",
           position: {
             boundingRect: {
-              width: 828,
-              height: 920,
-              x1: 9.2,
-              y1: 9.2,
-              x2: 18.4,
-              y2: 27.6,
+              height: 700,
+              width: 777.78,
+              x1: 155.56,
+              y1: 155.56,
+              x2: 194.44,
+              y2: 194.44,
             },
             rects: [],
             pageNumber: 1,
           },
-          redactionAddedOrder: 1,
+          redactionAddedOrder: 0,
           redactionType: { id: "1", name: "Named individual" },
           textContent:
             "This is an area redaction and redacted content is unavailable",
           type: "redaction" as "redaction" | "search",
         },
         {
-          highlightType: "linear" as "area" | "linear",
-          id: "1707153788147",
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
           position: {
             boundingRect: {
-              height: 920,
-              width: 1314.29,
-              x1: 131.43,
-              x2: 197.14,
-              y1: 131.43,
-              y2: 197.14,
+              height: 800,
+              width: 1100,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
             },
-            rects: [
-              {
-                height: 920,
-                width: 1314.29,
-                x1: 131.43,
-                x2: 197.14,
-                y1: 131.43,
-                y2: 197.14,
-              },
-              {
-                width: 1314.29,
-                height: 920,
-                x1: 144.57,
-                y1: 144.57,
-                x2: 210.29,
-                y2: 210.29,
-              },
-            ],
-            pageNumber: 1,
+            rects: [],
+            pageNumber: 2,
           },
-          redactionAddedOrder: 2,
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 800,
+              width: 800,
+              x1: 80,
+              y1: 80,
+              x2: 120,
+              y2: 120,
+            },
+            rects: [],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
           redactionType: { id: "1", name: "Named individual" },
           textContent:
             "This is an area redaction and redacted content is unavailable",
@@ -280,7 +254,6 @@ describe("redactionUtils", () => {
         expectedResult
       );
     });
-
     it("should return correct normalized highlight data, for highlight type linear", () => {
       const redactionHighlights = [
         {
@@ -288,8 +261,46 @@ describe("redactionUtils", () => {
           id: "1707153788147",
           position: {
             boundingRect: {
+              height: 800,
               width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [
+              {
+                width: 800,
+                height: 1000,
+                x1: 100,
+                y1: 100,
+                x2: 150,
+                y2: 150,
+              },
+              {
+                width: 800,
+                height: 1000,
+                x1: 110,
+                y1: 110,
+                x2: 160,
+                y2: 160,
+              },
+            ],
+            pageNumber: 1,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "linear" as "area" | "linear",
+          id: "1707153788148",
+          position: {
+            boundingRect: {
               height: 700,
+              width: 1000,
               x1: 100,
               y1: 100,
               x2: 150,
@@ -321,12 +332,101 @@ describe("redactionUtils", () => {
             "This is an area redaction and redacted content is unavailable",
           type: "redaction" as "redaction" | "search",
         },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 700,
+              width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "linear" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 900,
+              width: 1000,
+              x1: 200,
+              y1: 200,
+              x2: 250,
+              y2: 250,
+            },
+            rects: [
+              {
+                height: 900,
+                width: 1000,
+                x1: 200,
+                y1: 200,
+                x2: 250,
+                y2: 250,
+              },
+            ],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
       ];
 
       const expectedResult = [
         {
           highlightType: "linear" as "area" | "linear",
           id: "1707153788147",
+          position: {
+            boundingRect: {
+              height: 800,
+              width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [
+              {
+                width: 800,
+                height: 1000,
+                x1: 100,
+                y1: 100,
+                x2: 150,
+                y2: 150,
+              },
+              {
+                width: 800,
+                height: 1000,
+                x1: 110,
+                y1: 110,
+                x2: 160,
+                y2: 160,
+              },
+            ],
+            pageNumber: 1,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "linear" as "area" | "linear",
+          id: "1707153788148",
           position: {
             boundingRect: {
               width: 1142.86,
@@ -355,6 +455,57 @@ describe("redactionUtils", () => {
               },
             ],
             pageNumber: 1,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "area" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 700,
+              width: 1000,
+              x1: 100,
+              y1: 100,
+              x2: 150,
+              y2: 150,
+            },
+            rects: [],
+            pageNumber: 2,
+          },
+          redactionAddedOrder: 0,
+          redactionType: { id: "1", name: "Named individual" },
+          textContent:
+            "This is an area redaction and redacted content is unavailable",
+          type: "redaction" as "redaction" | "search",
+        },
+        {
+          highlightType: "linear" as "area" | "linear",
+          id: "1707153788149",
+          position: {
+            boundingRect: {
+              height: 700,
+              width: 777.78,
+              x1: 155.56,
+              y1: 155.56,
+              x2: 194.44,
+              y2: 194.44,
+            },
+            rects: [
+              {
+                height: 700,
+                width: 777.78,
+                x1: 155.56,
+                y1: 155.56,
+                x2: 194.44,
+                y2: 194.44,
+              },
+            ],
+            pageNumber: 2,
           },
           redactionAddedOrder: 0,
           redactionType: { id: "1", name: "Named individual" },
